Workaround CMake issues with 'install' binary
CMake gets really confused about a target and/or binary named "install".
We've already had a workaround in place, which is to name the target
"install-bin" and only set its executable name to "install". Unfortunately,
even that is not enough when using CMake with Ninja, as the Ninja target
still ends up being named "install" and so conflicting with the "install"
action.

This new workaround is pretty ugly, but it seems to work.